adjust import statments according to splitting-refactoing of nodes
