+++ src/bequem/nodes/__init__.py
@@ -1,16 +1,33 @@
 from bequem.nodes.node import Node as Node
 from bequem.nodes.proxy_node import ProxyNode as ProxyNode
-from bequem.nodes.constant import ConstantVector as ConstantVector, ConstantUnitary as ConstantUnitary
-from bequem.nodes.basic_ops import Tensor as Tensor, Adjoint as Adjoint, Scale as Scale
+
+from bequem.nodes.constants.constant_vector import ConstantVector as ConstantVector
+from bequem.nodes.constants.constant_unitary import ConstantUnitary as ConstantUnitary
+
+from bequem.nodes.basic_operations.tensor import Tensor as Tensor
+from bequem.nodes.basic_operations.adjoint import Adjoint as Adjoint
+from bequem.nodes.basic_operations.scale import Scale as Scale
+
+
 from bequem.nodes.permutation import Permutation as Permutation
-from bequem.nodes.controlled_ops import BlockDiagonal as BlockDiagonal
-from bequem.nodes.ring_ops import Mul as Mul, Add as Add
-from bequem.nodes.block_concatenation import BlockHorizontal as BlockHorizontal, BlockVertical as BlockVertical
+from bequem.nodes.controlled_operations.block_diagonal import BlockDiagonal as BlockDiagonal
+
+
+from bequem.nodes.ring_operations.mul import Mul as Mul
+from bequem.nodes.ring_operations.add import Add as Add
+
+
+from bequem.nodes.block_concatenation.block_horizontal import BlockHorizontal as BlockHorizontal
+from bequem.nodes.block_concatenation.block_vertical import BlockVertical as BlockVertical
+
+
 from bequem.nodes.identity import Identity as Identity
 from bequem.nodes.nonlinear import ComponentwiseMul as ComponentwiseMul
-from bequem.nodes.integer_arithmetic import (
-    Increment as Increment,
-    IntegerAddition as IntegerAddition,
-    ConstantIntegerAddition as ConstantIntegerAddition,
+
+
+from bequem.nodes.integer_arithmetic.constant_integer_addition import ConstantIntegerAddition as ConstantIntegerAddition
+from bequem.nodes.integer_arithmetic.constant_integer_multiplication import (
     ConstantIntegerMultiplication as ConstantIntegerMultiplication,
 )
+from bequem.nodes.integer_arithmetic.increment import Increment as Increment
+from bequem.nodes.integer_arithmetic.integer_addition import IntegerAddition as IntegerAddition
