+++ src/bequem/nodes/permutation.py
@@ -4,8 +4,9 @@ import numpy as np
 import tequila as tq
 from bequem.nodes.node import Node
 from bequem.nodes.proxy_node import ProxyNode
-from bequem.subspace import Subspace, ZeroQubit
-from bequem.nodes.basic_ops import UnsafeMul, Adjoint
+from bequem.subspace.subspace import Subspace, ZeroQubit
+from bequem.nodes.basic_operations.unsafe_multiplication import UnsafeMul
+from bequem.nodes.basic_operations.adjoint import Adjoint
 from bequem.circuit import Circuit
